crm_django/requirements.txt
asgiref==3.3.4
crispy-tailwind==0.5.0
Django==3.2
django-crispy-forms==1.11.2
django-environ==0.4.5
Pillow==8.2.0
pypng==0.0.20
PyQRCode==1.2.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
